Configurando o projeto para fazer rotas e criando rotas básicas para entender o conceito.
